Util functions for trigger and update; Emergency Temp Fix to Navbar Anchors
