+++ client/package.json
@@ -23,12 +23,14 @@
     "react-dom": "^19.0.0",
     "react-icons": "^5.4.0",
     "react-router-dom": "^7.1.1",
+    "react-router-hash-link": "^2.4.3",
     "react-swipeable": "^7.0.2"
   },
   "devDependencies": {
     "@types/bootstrap": "^5.2.10",
     "@types/react": "^19.0.2",
     "@types/react-dom": "^19.0.2",
+    "@types/react-router-hash-link": "^2.4.9",
     "@typescript-eslint/eslint-plugin": "^8.18.2",
     "@typescript-eslint/parser": "^8.18.2",
     "@vitejs/plugin-react": "^4.3.4",
